nyc_taxi_take_home_final/pandas_version/requirements.txt
pandas==1.5.2
azureml.opendatasets==1.55.0.post1